python/packages/kagent-sota-adapter/pyproject.toml
[project]
name = "kagent-sota-adapter"
version = "0.1.0"
description = "Generic CLI-to-A2A adapter for KAgent - wrap CLI-based AI agents as BYO agents"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "a2a-sdk>=0.3.23",
  "kagent-core>=0.1.0",
  "kagent-adk>=0.1.0",
  "httpx>=0.25.0",
  "fastapi>=0.100.0",
  "uvicorn>=0.20.0",
  "pydantic>=2.0.0",
  "typing-extensions>=4.0.0",
  "openai>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.0",
]

[tool.uv.sources]
kagent-core = {workspace = true}
kagent-adk = {workspace = true}

[build-system]
requires = ["setuptools>=61.0", "wheel>=0.46.2"]
build-backend = "setuptools.build_meta"

[tool.hatch.build.targets.wheel]
packages = ["src/kagent"]

[tool.ruff]
line-length = 120
target-version = "py310"
lint.select = ["E", "F", "I", "N", "W", "UP"]
lint.ignore = ["E501", "N803", "N806"]
